+++ dashboard/lib/service/data_seeder.dart
@@ -8,6 +8,7 @@ import 'dart:math' as math;
 import 'package:cocoon_common/task_status.dart';
 import 'package:cocoon_integration_test/cocoon_integration_test.dart';
 import 'package:cocoon_service/cocoon_service.dart';
+import 'package:fixnum/fixnum.dart';
 import 'package:github/github.dart';
 import 'package:googleapis/firestore/v1.dart' as g;
 import 'package:uuid/uuid.dart';
@@ -483,7 +484,7 @@ class DataSeeder {
       attemptNumber: attemptNumber,
       creationTime: creationTime,
       buildNumber: 1337 + attemptNumber,
-      buildId: 24567 + attemptNumber,
+      buildId: Int64(24567 + attemptNumber),
       summary: switch (status) {
         .succeeded =>
           '[INFO] Starting task $jobName...\n[SUCCESS] All tests passed (452/452)',
